% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_table.R
\name{create_dictionary_table}
\alias{create_dictionary_table}
\title{Create a dictionary table from YAML data.}
\usage{
create_dictionary_table(filename)
}
\arguments{
\item{filename}{YAML filename.}
}
\value{
Creates a markdown table.
}
\description{
Create a dictionary table from YAML data.
}
