% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary_entry}
\alias{dictionary_entry}
\title{Create a properly formatted data dictionary entry.}
\usage{
dictionary_entry(variable, unit, type = c("Numeric", "Discrete", "Date"),
  description)
}
\arguments{
\item{variable}{The name of the added variable (eg. 'BMI').}

\item{unit}{The unit of the variable (eg. 'kg')}

\item{type}{The data value type (eg. 'Numeric')}

\item{description}{A brief description of the dictionary entry.}
}
\description{
Can only be added to the data dictionary using \code{\link{dictionary_update}}.
}
