% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary_update}
\alias{dictionary_update}
\title{Update the dictionary with the entries created using \code{\link{dictionary_entry}}.}
\usage{
dictionary_update(entries, force_update = FALSE,
  dict_name = "added-variables.yaml")
}
\arguments{
\item{entries}{Properly formatted entries piped (or not) from the
\code{\link{dictionary_entry}} function.}

\item{force_update}{Force the dictionary file to be updated (ie. the entry is
removed and added again).}

\item{dict_name}{Name of the dictionary yaml file.}
}
\description{
Adds the entries to the \code{added-variables.yaml} file.
}

