% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_table.R
\name{generate_dictionary}
\alias{generate_dictionary}
\title{Generate the data dictionary tables of all yaml files.}
\usage{
generate_dictionary()
}
\value{
Prints all tables related to the dictionary.
}
\description{
This is used on the PROMISE data dictionary to create the dictionary from the
rename yaml files.
}
