% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{insert_formula}
\alias{insert_formula}
\title{Insert formula for created variables used in PROMISE.}
\usage{
insert_formula(variable = c("eGFR_f1", "eGFR_f2", "eGFR_m1", "eGFR_m2",
  "HOMA", "ISI", "IGI_IR", "ISI2"))
}
\arguments{
\item{variable}{The variable's formula to output.}
}
\value{
Outputs a string of the LaTeX math equation.
}
\description{
Insert formula for created variables used in PROMISE.
}

