% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_manuals.R
\docType{data}
\name{view_manual}
\alias{list_manuals}
\alias{view_manual}
\title{View the various PROMISE manuals.}
\format{An object of class \code{character} of length 2.}
\usage{
view_manual(manual = list_manuals, view_from = c("rstudio", "browser",
  "website"))

list_manuals
}
\arguments{
\item{manual}{The manual to view.}

\item{view_from}{Whether to view the manual from a browser or inside RStudio.}
}
\description{
These manualsthat contain more extended documentation on the data cleaning
and wrangling processes and on how to add to the dataset whenever new data
comes in.
}
\keyword{datasets}

